+++ pvar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for pvar
 
+## 0.1.2.0
+
+* Rename `modifyPVar` to `fetchModifyPVar` and `modifyPVarM` to `fetchModifyPVarM`. This
+  is a breaking change in favor of consistency with other librarries.
+* New implementation for `modifyPVar` and `modifyPVarM` that can return some artifact.
+* Addition of `modifyFetchPVar` and `modifyFetchPVarM`
+* Addition of `atomicModifyFetchIntPVar` and `atomicFetchModifyIntPVar`
+
 ## 0.1.1.0
 
 * Addition of backwards compatible: